PP-9333: frontend charge response includes nested `agreement`
Agreements are intentionally loosely coupled to charges to allow the
code that manages agreements to be moved if required. They are joined by
a guaranteed unique external identifier.

If an agreement ID exists on a given charge, lookkup the agreement and
include it in the frontend charge response builder.

Note this should be a very small percentage of general "WEB" payments
with the user present as it will only be the case when a user is both
paying and saving their card payment details to be used by a future
agreement.

Use the existing `AgreementResponse` serialiser to avoid including deep
embeds like the agreeement gateway account.