PP-9142 Change default value for authorisation_mode
To be consistent with other enum values, and to make serialization
and deserialization of the enum easier, the `authorisation_mode` should
be stored in the databased capitalized.

Drop the column and re-add with the correct default of "WEB". This is
possible as there are no charges currently in the database that have
been created with a different value.