PP-9142 Accept authorisation_mode when creating charge
Accept the field `authorisation_mode` in the POST request body to create
a charge. This can be any values for the AuthorisationMode enum in
pay-java-commons - `web`, `moto_api` or `external`. Save this on the
charge, and return in charge responses.

We only ever want the `external` authorisation_mode to be set internally
for payments created by the telephone payment notification API.
Validation to prevent this mode being used when creating a standard
payment through the API will be added in public API.