PP-9368 Extract code for marking charges as eligible for capture
Extract the code that marks a charge as being eligible for capture
into two separate classes: one that’s used for all charges and one
that’s used only for capturing delayed capture charges. The code
for this was previously split between CardCaptureService and
ChargeService.

There are absolutely no changes to funcionality of the code.

The changes to the tests are more substantial. Previously the unit
tests were ostensibly testing CardCaptureService but that
CardCaptureService used a real ChargeService, so some of the code
that was being exercised was actually in ChargeService (and some
relating to state transitions was actually in ChargeEntity!). The
new tests only exercise the specific code in the new extracted
classes and stub out ChargeService, with the assumption that any
methods being called on ChargeService and ChargeEntity are already
well-tested elsewhere.

This change is being made because we’re about to make marking a
charge as being eligible for capture more complicated when an
initial payment is made for an agreement and we want it not to be
a nightmare to implement and test.