PP-9368 Link payment instrument to agreement on user confirm
When a user confirms a payment, marking it eligible for capture
(CAPTURE APPROVED for most payments, AWAITING CAPTURE REQUEST for
delayed capture payments), link the payment instrument associated
with the payment to the agreement associated with the payment and
set the payment instrument to active if and only if the payment
has the flag to save the payment instrument to the agreement
enabled.

A @Transactional annotation makes sure that this linking and
status updating happens in the same transaction as the one that
updates the status of the payment itself, ensuring consistency.

The code accounts for three errors that can Never Happen™ because
they would mean the world has gotten into an inconsistent state:

1. The payment has no payment instrument on it (in which case the
   flag should not be set)
2. The payment has no agreement ID set on it (in which case the
   flag should not be set)
3. The payment has an agreement ID but there is no matching
   agreement in the database (in which case our foreign key
   constraints may have failed us)

If any of these errors occur, we log an error but continue marking
the payment as eligible for capture. This means that this payment
will succeed but the service will not be able to take future
recurring payments from the agreement.

Other errors (for example, the database connection failing), will
cause the entire process of marking the payment eligible for
capture to fail, which is consistent with other parts of the
system.