PP-9389 Fix NullPointerException
A NullPointerException was thrown when there is no return_url and the
account is live as we attempted to perform a validation on the return_url
before checking whether it was present.

Make the returnUrl getter on the CreateChargeRequest return an Optional
to force doing a check for whether it is present wherever it is used to
avoid this situation occurring.