PP-9503 Make the time unit of timeout configurations consistent
Make both the configuration options for the timeout for synchronous and
asynchronous payments use milliseconds as a unit for consistency.

Rename the environment variable AUTH_READ_TIMEOUT_SECONDS to
AUTH_READ_TIMEOUT_MILLISECONDS. It is fine to do this, as it is not set
to anything other than the default value in any of our infra.