PP-9503 Don't store updated gateway transaction id on ChargeEntity
When making an authorisation for Worldpay, if we request a 3DS exemption
and it is soft declined, we need to make the second authorisation
request with a new gateway transaction id (order code).

We were updating the gatewayTransactionId field on the ChargeEntity to
store this new id for retrieval when building the request to send to
Worldpay. This is confusing, as this updated value is never persisted
until after the Authorisation is complete, when it is extracted from
the BaseGatewayResponse.

Improve this by storing the gateway transaction id in a separate field
on the AuthorisationGatewayRequest. When we make the second
authorisation request in WorldpayPaymentProvider, construct a new
CardAuthorisationGatewayRequest from the original request and a new
gatewayTransactionId. This pattern will become more obvious when the
AuthorisationGatewayRequest no longer allows retrieval of the
ChargeEntity object.