PP-9503 Add ChargeEntity parameter to authorise method
Add a ChargeEntity parameter to the PaymentProvider.authorise method.

Previously this was passed inside the AuthorisationGatewayRequest.
However, we don't want the code that does the actual authorisation
HTTP request to have access to the ChargeEntity as this code should not
be able to persist any data on the charge.

The WorldpayPaymentProvider requires the ChargeEntity as it needs to do
some updates to charge using the result of a 3DS exemption request.

We are doing this because want to create a separate authorisation method
that doesn't need the ChargeEntity and does not attempt to persist
anything for use by the MOTO authorisation API code, which will run this
code in a separate thread with a timeout before attempting to update the
charge itself. This code will  still need to use the
AuthorisationGatewayRequest, so removing the  ChargeEntity from this
will make it easier for us to discourage introducing persistence to this
code path which could cause database conflicts.