PP-9503 Fix stubbing for unit test
Unit test was failing when run locally using IntelliJ even though it
passed on CI. This was because it was expecting the same
CardAuthorisationGatewayRequest to be used for both requests when a
Worldpay exemption was requested, but failed. However, a new
CardAuthorisationGatewayRequest is now created, so update mocking and
verification for tests to reflect this.