PP-9369 Allow "agreement_id" and "authorisation_mode": "agreement"
When creating a charge, update the validation to accept
"agreement_id" if accompanied by "authorisation_mode":
"agreement". Previously, "agreement_id" was only accepted if
accompanied by "save_payment_instrument_to_agreement": true (which
remains allowed). A request with "agreement_id",
"authorisation_mode": "agreement" and
"save_payment_instrument_to_agreement": true is not allowed.

This only affects the validation; there are no changes to how the
charge is processed once it is created. If a charge is created with
an "agreement_id" and "authorisation_mode": "agreement", it will be
treated the same as a regular charge with neither of those
properties. Only public API makes requests to the connector
endpoint altered by this commit, so that will not happen for now.

This commit also expands the test coverage for existing
functionality relating to creating charges associated with
agreements.

In order to reduce surprising behaviour, this commit changes how
default values work for ChargeCreateRequest (such as the default
authorisation mode being WEB). Previously, this was done by
having the fields assigned to a default value when they were
declared, which was unreliable because the tests can construct a
ChargeCreateRequest using a constructor that can cause a non-
primitive field to have its value explicitly set to null.
Instead, all fields are now initialised to null (including ones
that were previously primitives like the delayed capture boolean
flag — these have been boxed) and the default is returned in the
getter if necessary. This has the side-effect of allowing us to
make the toString() method more accurately reflect the original
request (e.g. by not pretending the delayed capture flag was set
to false when it was actually not included in the request at all).