PP-9369 Check agreement is valid before taking payment from it
Update the validation so that if a request is received with
"authorisation_mode": "AGREEMENT" and an "agreement_id", we check
that the agreement exists, has a payment instrument and the
payment instrument is in the correct ACTIVE state.

None of the logic after this validation has changed yet so if the
conditions above are true, we’ll just process it as a regular
user-present payment but that’s fine because public API will not
be sending us anything with "authorisation_mode": "AGREEMENT" yet.

If the agreement or payment instrument are not found, the error
returned is a 400 (bad request). While a 404 (not found) may seem
logical, the problem is that the request references a non-existent
resource, noot that the request is for a non-existent resource.
Making a request to save a payment instrument to a non-existent
agreement now also returns a 400 for consistency.

Based on code started by @jbz149 and @kundanscorpio