PP-9790 Make sandbox card numbers more flexible
SandboxCardNumbers is now an interface with two implementations:
SandboxFullCardNumbers and SandboxLast4DigitsCardNumbers
(formerly SandboxWalletCardNumbers).

Code in SandboxGatewayResponseGenerator (which is now a concrete
class rather than an interface implemented by
SandboxPaymentProvider and SandboxWalletAuthorisationHandler) can
use either implementation of SandboxCardNumbers to map between
a card number (full or last four digits) and a sandbox gateway
response.

The code in SandboxPaymentProvider (which knows what kind of
authorisation it’s dealing with it) can then choose how it wants
SandboxGatewayResponseGenerator to do the mapping.

This will be useful for authorising recurring card payments, where
we have a payment instrument with the last four digits but not a
full card number.