PP-9851: add operation key to charge operations queue message
- favour jackson object serialisation over building custom JSON maps as
  we were doing before, the object used to write to the queue message
  will be the same as the object used to read

Add an operation key attribute that will allow the queue handler to
perform different jobs based on the key. All jobs will be acting on
charges so the charge id attribute remains unchanged.

The current job is to `CAPTURE` a charge which is now set when
interacting with the queue.