PP-9791: associate charge with agreement payment instrument
When creating a charge with `authorisation_mode` `agreement`, associate
the created charge the agreements payment instrument.

There is already validation logic in place to ensure that the
appropriate combination of parameters is used and that a valid agreement
and payment instrument are in place - we can assume these are correct by
this point in the code.

Add an assertion to make sure we are correctly setting the payment
instrumnet.

Signed-off-by: Steven <steven.fountain@digital.cabinet-office.gov.uk>