PP-9791: move adding charge to auth job into service
Previously the charge ID was being added to the task queue using the
`offerTasksOnStateTransition` hook during state transitions.

We decided that this wasn't at all clear to someone reading the code,
especially if they are not familiar with this hook -- instead move this
call explicitly into the
`markChargeAsEligibleForUserNotPresentAuthorisation` method.

Clean up naming and test assertions.