PP-9677 Only send fee and net when dispute recharged
Only send the `fee` and `net_amount` fields in the DISPUTE_LOST event
when we have recharged the service for the dispute. This is because
these fields in our reports are used for reconcilliation, so should only
be included when money has moved into or out of the service's Stripe
account.