PP-9677 Process dispute transfers included in payouts
Handle Stripe BalanceTransactions which have a source of type
Transfer, and metadata "reason" field `transfer_dispute_amount` by
emitting a DISPUTE_INCLUDED_IN_PAYOUT event. The event_details for this
event contains the gateway_payout_id, which ledger will use to associate
the dispute with the payout.

When a Transfer has a "reason" metadata field that is unrecognised or
not handled, make the payout processing fail. This is to prevent us from
wrongly reconciling any future Transfer types we add as refunds before
we add the logic to the payout reconcile process to deal with them.

Use the MDC to add the payout id and connect account id to all log lines
for the payout being processed.