PP-9671 Return error when refund unavailable for disputed payment
When a request is made to refund a payment, and the refund status is
"unavailable" and the payment is disputed, return an error response with
identifier REFUND_NOT_AVAILABLE_DUE_TO_DISPUTE.

We only know whether a payment is disputed when we get the payment
from ledger. This means that if the payment is still in the connector
database, we will just return the standard error response with
identifier REFUND_NOT_AVAILABLE. Public API will handle this by
returning an error message stating that the refund is unavailable, but
will not be specific that it is due to the payment being disputed.

Update the pact test with ledger to expect the "disputed" field to be
present for a payment.