PP-9671 Emit refund availability updated event when dispute won
When we receive a notification from Stripe informing us that a dispute
was won, emit a REFUND_AVAILABILITY_UPDATED event to ledger to update
the refund status.

We do this as we will update the refund status to "unavailable" when
a new dispute is raised. When the dispute is won be the service, the
availability should then be reset to the status it would be if there
was no dispute.

Fix some use of optionals in EventFactory.java while refactoring to
re-use some code.