PP-9671 Add magic cardholder name to skip submitting dispute evidence
Stripe provide 2 test card numbers which will create a dispute as soon
as they are captured. On receiving the dispute created webhook, we
automatically submit either winning or losing evidence to Stripe for
these card numbers.

So that we can internally test the state of a payment when the
dispute is still open, add a magic cardholder name: 'skip_evidence',
which when used will cause evidence not to be automatically uploaded.
This is necessary as there is not a different card number we can use to
create the dispute.