PP-9906 Better handle retying capture for Stripe
A capture can fail if Stripe returns an error response to either the
request to capture the payment intent, or the request to transfer the
funds for the payment to the connect account. If just the transfer fails
we are often able to recover from this due to the idempotency of
Stripe's API.

However, if Stripe returns an error response for the capture attempt but
it actually succeeded on Stripe's end then Stripe will continue to
return an error response on retry attempts until the idempotency expires,
after which time it will return a different error stating that the
payment intent has already been captured.

Make changes to the capture process so that if we are trying to capture
a payment where capture has previously been attempted, first query
Stripe to find out whether the capture was already successful.

If Stripe says the payment has been captured, skip making the capture
request. Then check whether a transfer already exists, and only attempt
the transfer if no transfer exists.