PP-9929: make payment instrument card type nullable
Card type is nullable on charges, currently the enum serialisation for
card types uses null to represent that card id didn't know if the card
number was credit or debit (`CREDIT_OR_DEBIT` -> `null`).

There should be exact parity between the current details stored on
charges and payment instruments to give confidence that running shared
code will work.

Adds a test using a card number with type `CREDIT_OR_DEBIT` that would
fail unless the `card_type` property on payment instruments has been
made nullable.

If we've moved over to payment instruments entirely in the future this
behaviour can be changed but for now it should be consistent with
charges.