PP-7862 Respond to notification with 500 when ledger returns error
When attempting to get a transaction from ledger and ledger returns a
response other than 200 - OK or 404 - NOT FOUND, throw an
exception and log an error. When ledger returns a non-success code other
that 404 that indicates an issue that should be investigated.

Throwing a LedgerException will cause the HTTP request to connector to
respond with a 500 status code. This includes for the HTTP request from
gateways. In this case, we want to return a 500 response so Worldpay
will retry sending the notification at a later time.

The other places where connector attempts to get transactions from
ledger is for parity checking and for handling Stripe webhooks. In both
these places, returning a 500 when ledger returns an error is also
sensible behaviour.