PP-10173 Use Instant not ZonedDateTime for timestamp in Event
Make the timestamp field in Event (the superclass of the events
that get emitted to ledger) be an Instant rather than a
ZonedDateTime.

This does not change how a subclass of Event is serialised to
JSON.

Keep the existing constructors that use ZonedDateTime so we can
migrate the subclasses gradually.

This may change how the timestamp is stringified in the toString()
method for subclasses of Event in some circumstances but not in a
significant way.

To avoid NullPointerExceptions triggered by some new conversions
from ZonedDateTime to Instant, some tests were updated to include
a timestamp in an Event where there previously was not one. This
is fine as in reality all events have timestamps (except for
UnspecifiedEvent).