PP-8693 Return 402 for gateway errors
Connector currently returns 500 when there are gateway errors.
When front end receives 500 from connector this implies the problem
is with connector rather than with the PSP, and causes unnecessary alerts.
Instead 402 will be returned when there are gateway errors.
A change will be made to the front end so that it handles 402.
Front end will display the same error page to the user as it does for 500.