PP-10435: use appropriate state for capturing user not present payments
The default state for marking a charge as eligible for capture moves the
charge into the `CAPTURE_APPROVED` state, this transition is annotated
by the `UserApprovedForCapture` event.

As there are no users directly involved in recurring (agreement)
payments, update the method to determine what state to move to to also
check for agreement authorisation mode. The `CAPTURE_QUEUED` state
indicates the charge will be captured but instead uses the
`QUEUED_FOR_CAPTURE` event which doesn't refer to a paying user (rather
a system call).