PP-10542: Process balance.available Stripe notification (#4232)
* PP-10542: Process balance.available Stripe notification

It is useful for product owners to see the length of time our Stripe balance is in the
negative. In order to do this we've enabled the balanace.available events
webhooks from Stripe. Upon receiving this event, log the currency and amount
which splunk can use to produce a nice graph of our Stripe balance over time.