PP-10427 Remove separate OrderRequestType
We were going to use a separate OrderRequestType for gateway requests
that authorise a payment and store payment details. This would have
given us separate metrics.

This is going to be re-visited in a separate story just to look at the
metrics, so for now just use OrderRequestType.AUTHORISE for operations
that set up recurring payment agreements.