PP-10430 Fix null pointer exception
It is possible for the CardType saved on a PaymentInstrumentEntity to
be null.

Handle this when we convert the PaymentInstrumentEntity into an
AuthCardDetails by treating null as PayersCardType.CREDIT_OR_DEBIT, as
this is what we do when we convert from a CardidCardType to
a PayersCardType for taking user present payments.