PP-10513 Refactor DeleteStoredPaymentDetailsGatewayRequest
Don't implement the GatewayRequest interface. This interface isn't
directly referenced anywhere so there is no need to implement it, it
seems it was just created for code organisation purposes. There is no
AuthorisationMode associated with this request, so it doesn't make
sense implementing the getAuthorisationMode interface method.

Have getters for the individual properties of the Agreement, Payment
Instrument and Gateway Account. The intention of the GatewayRequest
classes is to pass information to the PaymentProvider code level.
It looks like we try to only allow access to the individual
bits of information that the PaymentProvider code needs rather than the
entire database entities. This is likely to make sure that the
PaymentProvider code does not modify the database entities, as this is
outside of its concern.